Create textbook in progress. Main issue comes from incompatability
Create doesn't work due to a incompatability with the datepicker widget? I think it only accepts DateField, while our model has DateTimeField, but it doesn't really makes sense to have the publication date as DateTimeField since we don;'t need the time lol. The bigger issue that comes from this is that we might have to fix our fixtures that have the publication dates